Port System Settings to Interview using a copy of Dolphin's KCategorizedView.  This improves item layouting now and will allow more appropriate sorts, as well as alternate views (KControl emulation mode anyone?).
This replaces the attractive desaturate effect for search misses with 
the standard behaviour that only hits are shown by the proxy model.
This is arguably less attractive but more familiar and accessible.

The KCategor* classes are COPIED from Dolphin but will be synced
and removed when they move to kdelibs in 4.1.

TODO: figure out why KCategorizedSortFilterProxyModel doesn't emit
layoutChanged() when the model is filtered - this prevents us updating
the number of search hits at present.
TODO: get ereslibre to remove the omnipresent rubberband from
KCategorizedView
TODO: add weights to the categories and sort the model according to them
TODO: add a custom KCategoryDrawer to draw the category icon in the view
TODO: add a custom delegate?


svn path=/trunk/KDE/kdebase/workspace/; revision=752203