Honour the NoDisplay attribute of KServices
Summary:
kcmshell5 refuses to load modules that have this set and
when clicking on them in systemsettings, an empty page with
no button to go back to the overview is shown. So just don't
show them to the user.

Test Plan:
Started systemsettings5 in a X11 session, "Accessibility" is shown.
Did the same in a wayland session, it is not shown anymore.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6612