Add hidden system setting group
Summary:
We have a KCM (my plasma renderer settings) which we want to hide in
system settings, but still findable in krunner manual kcmshell5
invocation. Setting the relevant .desktop to hidden breaks the latter.

This patch creates a group which is hidden. KCMs parented to a hidden
group remain hidden when shown in a grouped way.

Test Plan: Opened system settings

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12449