[Icon View] Don't use dialog-warning emblem on icon when KCM requires authentication
Summary:
Icon view currently sets a dialog-warning emblem on the sidebar icons for KCMs that require root authentication (SDDM, KDE Wallet).

This overlay is both confusing and unnecessary given that we already spell it out in big letters in the KCM header that we'll be asking for authentication.

This patch removes the icon overlay.

BUG: 413183

Test Plan:
Before:
{F7737211}

After:
{F7737213}

Reviewers: #vdg, ngraham, ndavis

Reviewed By: #vdg, ngraham, ndavis

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25150