[Icon View] Fix titles of previously opened KCMs bleeding into QML KCMs
Summary:
The change in D14580 triggered the spawning of an additional header in QML KCMs in icon view mode.

The additional header was the header of the previously opened QWidgets KCM.

Sidebar view doesn't have this issue so I went and found the code that prevents the bug from happening and added it to icon view.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25726