Fix Ctrl+F shortcut conflict
Summary:
The base app was always setting Ctrl+F as a search shortcut, even when the toolbar upon
which the search field lives was not visible. This caused the dreaded "ambiguous shortcut"
dialog to appear. The patch fixes the issue by setting the shortcut only when the toolbar
contains a search field, and unsetting it when the toolbar is hidden.

BUG: 405030
FIXED-IN: 5.18

Test Plan:
Hit {key Ctrl F} in both sidebar and icons view. In both cases, it focuses the search
field.

Reviewers: #plasma, GB_2, apol

Reviewed By: apol

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26305