Add global shortcut
Summary:
The "Tools" key is function key in some laptops. In ThinkPad + Windows, when you press that key, it will open Control Panel or Windows 10 Settings. But in Linux, it does nothing. So I think it can be a good shortcut for System Settings because no other software use it.

{F8309275}

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: ngraham, davidre, ognarb, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29625