[sidebar view] Remove size hacks for toolbar items
The heights of these items are all bound together to keep them the same
height, but this is the wrong way to do it; instead, the QQC2 styling
for these controls should automatically take care of ensuring consistent
heights, as I recently did for the PlasmaComponents3 versions in
https://invent.kde.org/frameworks/plasma-framework/-/commit/e607c1ae470acd526c8a0a0d8323e878e8be55e8.

These hacks have caused a bug when the font size is increased. Let's remove
the size hacks and fix the sizing properly in the style itself.

BUG: 424228
FIXED-IN: 5.20.0