[sidebar view] Show list item separator lines
Showing lines is the default setting, so not doing so makes the sidebar
list visually inconsistent with the various other lists inside QML-based
KCMs.

My sense is that the reason why we hid the lines in the past was
because the lines had a very heavy appearance, but that was improved
recently to look much more lightweight.

Therefore, let's turn the lines on.