Rework main.cpp
The main change is that all commandline parsing happens before the KDBusService part is reached. This allows to unify the command line handling and error checking for the cases of launching a new systemsettings and activating an existing instance.

This fixes the issues that it was impossible to use systemsettings5 --list when an instance was already running and that it was impossible to pass arguments with a module when activating an existing instance. It also adds some additional error checking/reporting.

The SystemSettingsApp class and its (apparently unused DBus interface) is removed in the process.