[Sidebar view] Port headers to use Kirigami.AbstractApplicationHeader
This is the correct thing to do in general because custom views that
replicate standard components are bad. It also ensures that the headers
will change their colors properly as the color scheme changes once we
merge the Tools Area work.

The only Visual/UX change is that the subcategory header is made more
standard in appearance and behavior, with a discrete back button instead
of the entire area being clickable. This represents a mild regression in
usability since the clickable area is smaller now, but it's a win for
consistency and code simplicity.