[Sidebar view] Re-add mouse area for handling clicks on subcategory header
Now, even though there is a real button you can click on, if you happen
to click on any other part of the subcategory page header, it also goes
back, just like it used to. This is kind of hidden, but if we don't do
this, we will surely get complaints from people who had gotten used to
the nice big click area we had before.