RFC: Test GLContext creation at startup of system settings
The first frame safety method provides 2 functions:

A recent regression means that on Nvidia, after suspend, the driver
reports that the context is lost even in newly spawned applications
after the suspend.

This messes with QQuickWidget.

Querying it once (done by makeCurrent) will clear that flag and
everything behaves afterwards.

BUG: 424592

In addition we can detect the common bug of libGL being broken.

This commonly happens with a certain driver
updates, but requires running against a specific kernel version that is
not yet loaded. If we hit this, we respawn in software rendering mode.

BUG: 426019
(among 70 other duplicates all closed as not a bug or upstream)