Delete icons view
The icons view has seen substantial bit-rot and bugs due to not being the default view.
Additionally, it's missing out on many new features, such as quick settings and viewing
settings that have changed from the defaults.

Continuing to ship the icons view encourages some distros to ship software that nobody
is interested in maintaining, causing their users to receive a subpar and buggy experience
without the newest features we are working on.

Removing the icon view encourages distros shipping unmaintained software as the default
to switch to something that we are actually working on and putting new features and bugfixes
into.

The icons view is also somewhat incompatible with many design decisions we've made for the
arrangement of system settings, e.g. the appearance category does not look good being the
only icon in its section w/out a header. We cannot resolve this without going back on our
design to accomodate a non-default view used by a small proportion of users.

In all, icons view is a net negative to the experience, as we're currently shipping something
nobody cares to fix bugs on which increases the perception of our software being filled with
obscure bit-rotting settings in various nooks and crannies and which is not conducive to us
being able to ship more features and bugfixes for *everyone* in system settings.

Therefore, removing it is the best course of action here.

Due to system setting's modular design, users and distros can simply re-add the icons view out of
tree easily if they disagree with our decision to remove unmaintained software from our repos
and default offerings.