Port loading of categories away from KServiceTypeTrader
Instead load them from a specific folder in the generic data location.

This way it is an implementation detail if the MenuItem was created from
a config file, a service or in the future KPluginMetaData.
To provide access to the required properties simple getters are created.
This makes the API also more readable than doing file name checks somewhere
else in the code on a property that is optional.

Task: https://phabricator.kde.org/T14517