Port ExternalAppModule away from KCModuleInfo
The KCModuleInfo::service method is marked to be removed in KF6, also
KCModuleInfo should be removed.

Considering that we just need the service to read the name and to
pass it to KIO::ApplicationLauncherJob we can use a KService directly.

This will also make it easier to port the external app module loading away from KServiceTypeTrader.