Remove About Current Module
systemsettings has the ability to display a KAboutDialog for the current
module.

However that is quite hidden since it only works with the non-default
icon view or a global menu.

The fact that nobody has added it to the default sidebar view is a hint
for how useful it is percieved.

The information the dialog provides is not terribly useful IMHO and
often a bit broken since lots of KCMs either set no description or
authors in their KAboutData