Import services runner & strip out unneeded parts
This way the KCMs and applications will be provided by different runners.
When the KCM loading/querying is ported away from KServiceTypeTrader this is going to
make porting the runner far easier.
This will also be the case when a new job for launching KCMs is introduced - which
should not depend the desktop files but JSON metadata of the binary plugins.

It also allows users to disable the systemsettings results by disabeling the plugin.
This could be useful if one prefers to search for KCMs only inside of systemsettings.
Currently that could be done using a obscure setting in the KRunner config file, but the mechanism
behind it is deprecated for quite some time now and is not exposed in any GUI since KDE4.

It might also be worth noting that it is not user friendly to have the runner which provides the
"Systemsettings" and "System Information" categories under the name "Applications".