Don't request context help titlebar button
Most UI elements of most KCMs don't have What's This text set, so when
the user clicks on the button and hovers over UI elements, almost all of
them display the blocked cursor rather than showing the option for
context help. This makes the feature not very useful. So let's not opt
into it.