don't duplicate the logic for externalapp checking
this is already done inside the menuitem, no sense duplicating it.

addModule now consumes the item instead of the moduleinfo, inside it
then extracts the moduleinfo for ease of access. this then allows
addModule to call the existing isExternalAppModule() of the menuitem
thereby moving the specifics of what constitutes an external app out of
the moduleview.