add support for InfoCenterExternalApp
this entirely mirrors the existing SystemSettingsExternalApp tech.
unfortunately SystemSettingsExternalApp is a tad impossible to extend to
cover kinfocenter as well because the original lookup constraint is
simply that the service type be SystemSettingsExternalApp. ParentApp, or
indeed a lack of the key, did not factor into the lookup and as such
it's not really trivial to expand without going wild on the query
constraint (which seems ill advised considering we want to get rid of
servicetypetrader queries in general)

with this feature its now possible to also add externalapp modules to
kinfocenter by installing a suitable desktop file with
`X-KDE-ServiceTypes=InfoCenterExternalApp`

the motivation behind this is that we'd like to help users get from
kinfocenter to plasma-systemmontior if they find that what they actually
need is the current performance metrics rather than general purpose
system info from KIC, and external app should neatly solve this scenario