Create versionless systemsettings executable & install symlink to old name
This way consumers will be able to call "systemsettings" as the executable name
and don't have to worry about the major version. Considering the systemsettings is part of Plasma
and Plasma 5/6 are not going to be co-installable there is no point in having a versioned executable.

In KF6 the symlink to systemsettings5 can be removed.

Task: https://phabricator.kde.org/T14763