Do not focus the searchfield on startup if a virtual keyboard is enabled
If a virtual keyboard (like Maliit) is enabled, we end up always showing
the virtual keyboard on startup, while there are many cases where a user
would not use the search. So make use of new API added to Kirigami to
detect if a virtual keyboard is enabled and only focus the search field
if there is no virtual keyboard enabled.