Do not focus the searchfield on startup if a virtual keyboard is enabled
If a virtual keyboard (like Maliit) is enabled, we end up always showing
the virtual keyboard on startup, while there are many cases where a user
would not use the search. So make use of new API added to Kirigami to
detect if a virtual keyboard is enabled and only focus the search field
if there is no virtual keyboard enabled.

(cherry picked from commit 072883ae6bc5d8c350218927250a9a2b1b8e95c4)

Note that this commit was adjusted to check for the existence of the
InputMethod singleton to ensure we don't break things when running with
Kirigami 5.90.