sidebar: Make footer toolbar match height of app's own footer button row
System Settings internally adds a footer of buttons (Apply, Help,
Defaults, etc) to all KCMs. The height of this toolbar was always taller
than the height of the footer toolbar for the sidebar view. As a result,
whey did not have the same metrics and their heights differed. This was
never a problem before https://invent.kde.org/frameworks/kdeclarative/-/merge_requests/91
because they were never trying to be pixel-perfect with one another
before. But now they are, so this commit makes the sidebar's footer
match the height of System Settings' own footer.