sidebar: switch subcategory header's DIY back button to be a ToolButton
This is a more conventional UI element that fits just fine here. Due to
a limitation of how QQC2.ToolButton aligns content when forced to be
wide (see https://bugs.kde.org/show_bug.cgi?id=445033), we have to
create our own custom content item for it. It looks fine with all
tested color schemes as well as RTL mode.