Don't show unnecessary root authentication message
This message is pointless because the user will find out that they need
to authenticate when they click the "Apply" button anyway. Showing the
user this information ahead of time just contributes to visual clutter
when it is even shown.

However the message hasn't even been shown for at least 9 months due to
an intentional change in header height calculation to make the KCM
header height always match the height of the category/subcategory
sidebar header height. This had the side effect of entirely or mostly
cutting off the authentication message. With the current design, there
is no way to fix this without regressing the header alignment fix, and
the message is pretty much unnecessary anyway, so the simplest solution
is to not show it at all.

BUG: 434113
FIXED-IN: 5.24