FooterToolbar: Match also padding to the modules' buttons
Set the padding in the sidebar footer to match the value of the
breeze default padding (6px), so the "Highlight Changes" button
is consistent with the footer buttons in the modules.

The value is hardcoded for now, until we can get the margin values
from the QStyle (the same existent comment applies).

| BEFORE |![footer_before](/uploads/7ccc28ad71f29eca36a9a5a745c03eb6/footer_before.png)|
| ---|---|
| AFTER |![Screenshot_20220122_145735](/uploads/c879f898cb1651336da24a89c675c779/Screenshot_20220122_145735.png)|

@teams/vdg