ModuleView: Correctly update button box visibility
When buttons are hidden externally use a button mask to avoid having
visibility conflicts between the parent button box and the individual
buttons, since visible property is inherited.

A small refactor is required to de-couple an updateButtons() method.

Also remove the ability to individually hide the reset button, as this
is not defined as a module button, but attached to the Apply button.

(cherry picked from commit 224cccebb93088ae2c4579b1c290d675b6a923ff)