IconView: Fix header padding
We weren't setting the vertical spacing for the page's GridLayout,
so it took the standard 6px causing this extra padding on top and
bottom for every KCM.

Better apply this margins to the title widget which is only shown
for iconview.

The horizontal spacing is not needed, so that part can be simplified.

BUG: 449040
FIXED-IN: 5.24


(cherry picked from commit ed64dbc76457bc66eaf565bf57ea0bf379cf4c59)