De-duplicate KCMs based on plugin id of converted KPluginMetaData object
In case of KIO, the KCM metadata files need to be kept for compatibility, but they
reference the KCM in a proper Plasma namespace. However we need to set the plugin id explicitly
in order for the desktop file to be found then the KCM should be pinned.
The explicit value differs from the one we would derive from the library name.
This means that de-duplicating based on the library name no longer works.