IconMode: Load all of the KCMs in their category
When opening systemsettings with a specific KCM, it currently
just loads that KCM and the first one in the category.

Load instead all of the KCMs in the category.

This makes it consistent with the regular use of icon mode,
where the user navigates through full categories, and also
with the sidebar mode.

A different behavior is just confusing.