Remove window size limit
This was done years ago when very few KCMs had been ported to QML and
they reacted very badly to being made scrollable because the window was
too small, so we wanted to explicitly force the window to be quite big.
However things are much better now, and I think it would be safe to
reduce the minimum size and make the current minimum size be a size hint.