Don't show tooltips for Sidebar mode list items
System Settings currently shows large detailed tooltips when you hover
over a list item in System Settings' Sidebar mode, whether it be a
category or a KCM. If the list item is a KCM, the tooltip shows its
icon, name, and description. If it's a group, the tooltip shows you the
KCMs in that group, along with their names and icons.

This is inconsistent with all other multi-level list views and sidebars
in all other QtQuick apps (and probably QtWidgets apps too), and can
contribute to the "tooltip overload" effect sometimes seen in KDE apps,
whereby tooltips are popping up and disappearing constantly as you simply
move the cursor around.

In Icon mode, there is a bit of utility here to communicate to the user
when an icon is a group that contains KCMs, or simply a KCM itself. But
in Sidebar mode, this is visually communicated by the presence of arrows
for list items that contain groups of KCMs.

So let's not show the tooltips, but only for Sidebar mode. This removes
a ton of code, and eliminates an undesirable instance of code duplication
for the TooltipManager class.