Sidebar: Move final trailing separator out of main layout
tl;dr The scroll bars must not be overlayed with foreign items like
frame borders, separators etc. They are doing fine on their own. Don't.

And since Kirigami.ApplicationItem was not designed for this kind of
usage where you stick something permanent to one side, just more it out
completely in its own layout. Components are refactored in different
files, so that semantics of ApplicationItem being a QML context on its
own is preserved, and properties, methods & signals are wired together
to act as a proxy.