Cancel pending resolveChanges dialog when a new one is started
We run a nested event loop while the message dialog is shown and
is waiting for user input. When during this a new request to
open another module happened (via dbus) we would show a second
dialog and run another event loop. After the first dialog was
answered with discard changes, the module was destroyed, so the
second discard answer chrashes trying to reset module which was
no longer around.
Instead prevent this by rejecting the first dialog (effect being
staying on the current module and doing nothing) before showing
the new one. This also ensures that afterwards the correct module
is opened (the one from the second request).
BUG:465510
FIXED-IN:5.27.4

(cherry picked from commit 58907166d6600de3ad654922689af2a6bd0e1c51)