Allow updating arguments to the current module
When systemsettings is called on the current module,
we want to send an activation request instead of the
previous no-op.

This allows updating the call arguments on subsequent
calls. The KCM then can process these new arguments
as appropriate.

This is useful when we are calling systemsettings from
config buttons that point to a specific page or item