Remove soft hyphens in keywords used for searching
Some languages/translations use soft hyphens in long words
so that the words can be automatically split across lines.
The soft hyphen characters are normally invisible, and the
user never writes soft hyphens, so when searching for keywords
containg these characters, the characters have to be removed
in order for the search string to match the keywords.