Implement majority of settings re-organization
Part of https://invent.kde.org/plasma/systemsettings/-/issues/15.

This commit re-organizes categories to match what was agreed to in the
above-mentioned Issue.

It does not fully implement everything there because we're still missing a
few technical pieces that are blockers for fully completing the project:
- No infrastructure for embedding KCMs in other KCMs. So in places where
  we envisioned embedding, a status-quo-style child category is used
  instead. Not quite as nice, but close enough for now.
- No ability to hide KCMs (or categories) based on hardware
  availability, so more KCMs appear in the "Input & Output" category
  than were originally envisioned.
- The Date & Time KCM doesn't build against Qt6 yet, so it doesn't
  appear in the "Regional Settings" category as expcted.
- No "mouse and touchpad" icon, so the Mouse and Touchpad KCMs are still
  ungrouped.
- Kept the "Workspace Behavior" category name for now instead of
  renaming it to "Search & Shortcuts" as envisiones, because the other
  KCMs besides Search and Shortcuts that currently live in that category
  cannot yet be removed from it in the absence of the embedding feature
  or a better logical home for them.

Still, this MR plus the invididual ones in other repos amount to most
of the project from a UX perspective, and the remaining pieces can
trickle in over time and preserve the overall structure implemented
here.