categories: reverse location of "Internet" and "Network" text
The design goal here was to get the word "Internet" shown prominently,
but the way we did it for Plasma 6 introduces a logical categorization
error, because internet is a subset of networking, not the reverse.

Switching those around doesn't make the key word "Internet" disappear,
so I think there's no harm in correcting the categorization error.