Stop hiding homepage from sidebar, and remove "go home" toolbar button
Part of https://invent.kde.org/teams/vdg/issues/-/issues/51.

Per the HIG, when an app has a navigation control like a sidebar, the
page that the app opens to should be visible in it.

We already do this in most KDE apps (including Into Center, which shares
shell code) but System Settings was an odd exception. This commit
remedies that, and removes a whole lot of now-unnecessary supporting
code.

Info Center is unaffected.