Fix sub-category view remaining visible when moving to a top-level KCM
This fixes an issue where the sub-category view would still be visible
when opening a different kcm that is at the top level.

For example, if you did:
- systemsettings kcm_mouse
- and then, while System Settings is still open on that page did:
- systemsettings kcm_printer_manager

The sub category view on the left would still be on the Mouse & Touchpad
page. This commit fixes that by resetting the pageStack index to 0.

BUG: 434345
FIXED-IN: 6.1.2