Force focus to subcategory page only when actually needed
Previously we were forcing focus to it anytime it existed, even if it
did not have enough items to become visible. This bug was masked by the
previous code, which popped the subcategory column explicitly when
showing the intro page. Removing that in
ed05004b033950a4715ca7f0085198d2bed5b895 exposed the bug, which is fixed
in this commit.

BUG: 489209