ModuleView: only give ScrollArea focus if its useful
ModuleView places the loaded kcm in a QScrollArea so that it becomes
scrollable in case it is larger than the window size. This QScrollArea
takes keyboard focus. This is necessary to allow the user to pan the
view useing the keyboard, but is a completely useless tab focus chain
entry in the majority of cases, namely when either a qml kcm is loaded
(where the scrollability is usually handled with a ScrollView, as in the
components from KCMUtils), and whenever the viewport is large enough to
fit the whole kcm. To the user, this feels like focus vanishing into
thin air, as it lands on an element that cannot be interacted with at
all.

This change sets the QScrollArea to NoFocus if the content fits, and
only allows tab focus if it does not.