kcmmetadatahelpers: Filter out kcms by form factor
This commit introduces code to filter out a KCM if it is not of the same
form factor (platform) that Plasma has. This allows us to introduce KCMs
specific to mobile in packages shared between mobile and desktop (ex.
plasma-nm). I want to eventually consolidate some of the mobile KCMs
such as Wi-Fi and cellular data into plasma-nm so that there can be more
code sharing.

This has a caveat at the moment, on desktop
KRuntimePlatform::runtimePlatform() currently returns an empty list...
I'm not sure if this was intentional (I have a hack for it in the code
right now), but perhaps we need something like
https://invent.kde.org/plasma/plasma-mobile/-/blob/master/bin/startplasmamobile.in?ref_type=heads#L18
for Plasma Desktop?