SideBarItem: disable SubCategoryPage while invisible
The offscreen subCategoryPage while in single-column mode can in some
edge cases still receive keyboard focus, especially when the broken
keyboard focus in kcmutils/kcmoduleqml is replace with one that works
better. I have not managed to trigger the issue while being consciously
aware of it with the master version of kcmutils, but it's not unlikely
that it could happen there as well given the fragile and inconsistent
nature of current keyboard focus passing. If it does happen, focus ends
up stuck as it can't get itself into the visible parts anymore.
The invisible hamburger menu can even be activated.

So it makes sense to disable the page while it's not visible, and cut
the problem off at the root.