SubCategoryPage: don't push existing sidebar content
There are several places where sidebar is updated. One of these is in
the SubCategoryPage, which always pushes the current column into the
pageStack if the activeSubCategoryRow changes– even if the column is
already in the stack, which causes Kirigami to complain loudly.

Instead, we just don't push the column here at all; if a module is
loaded it will automatically do it anyway, and if the module does not
change there's no reason to load a new column anyway.