Support isRelevant and auxiliaryAction from KCModuleData
Filter KCMs based on their relevant property, e.g. whether the
device type, such as touchscreen or graphics tablet, is available
in the first place. A menu option to show all settings is added.
They are also shown when searching.

A tool button or a switch (in case of a checkable action) is added
based on the auxiliary action provided by the KCM.

To reduce plug-in loading overhead, the KCModuleDataV2 is only
loaded when the KCM announces X-KDE-System-Settings-Uses-ModuleData