Retranslate QML on QEvent::LanguageChange
Kirigami loads its qm file with the automagic ecm loader.

Since systemsettings does not link to kirigami, kirigami is loaded by
QML which happens to be in a thread, which the ecm loader says "i need
to load the translator in the main thread", which is too late.

We could also fix this linking the app to Kirigami but doing it the
generic way means it will work for any other such case too.

BUGS: 512187